requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==0.5.0
Django==3.2.6
django-admin-sortable2==1.0
django-appconf==1.0.4
django-compressor==2.4.1
django-daterangefilter==1.0.0
django-extensions==3.1.3
django-js-asset==1.2.2
django-mptt==0.13.1
django-photologue==3.13
django-recaptcha==2.0.6
django-robots==4.0
django-sortedm2m==3.0.2
django-tinymce4-lite==1.8.0
django-toggle-switch-widget==0.1.1
django-trumbowyg==1.1.2
django-watson==1.5.5
ExifRead==2.3.2
jsmin==2.2.2
kombu==5.1.0
Pillow==8.3.1
prompt-toolkit==3.0.20
psycopg2==2.9.1
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
